refactor: rename agent_memory.py to agent_history.py, rename test_agent_memory.py to test_agent_history #115
