+++ atomic-examples/web-search-agent/web_search_agent/main.py
@@ -5,7 +5,7 @@ from rich.markdown import Markdown
 from pydantic import Field
 
 from atomic_agents.agents.base_agent import BaseIOSchema
-from atomic_agents.lib.components.agent_memory import AgentMemory
+from atomic_agents.lib.components.agent_history import AgentMemory
 from atomic_agents.lib.components.system_prompt_generator import SystemPromptContextProviderBase
 
 from web_search_agent.tools.searxng_search import (
